tests(wot): IdentityIndexOf it's correctly pruned now
Since commit 012d096f31c6754deb8a7f5626fb4aab3d85f19a,  IdentityIndexOf it's correctly pruned. Before this commit, we found the associated author, even for a removed identity, it still failed when we tried to modify the identity, but then the error returned was not the right one.