docker: update docker-compose files
Using websocket address form (tcp/30333/ws) for the RPC node and TCP
form (tcp/30333) for the validator node, because as explained in the
node-template / substrate documentation [1], the default listen
addresses are:
* /ip4/0.0.0.0/tcp/<port> and /ip6/[::]/tcp/<port> when using --validator
* /ip4/0.0.0.0/tcp/<port>/ws and /ip6/[::]/tcp/<port>/ws otherwise.

[1] https://docs.substrate.io/reference/command-line-tools/node-template/

The argument of the '--public-addr' duniter option should not feature the
p2p endpoint part '/p2p/<PEER_ID>'. Duniter adds it automatically.

For example, when using:
  --public-addr /dns/gdev.example.com/tcp/30333/ws
you'll see this in the logs:
  Discovered new external address for our node:
    /dns/gdev.example.com/tcp/30333/ws/p2p/<PEER_ID>

When using:
  --public-addr /dns/gdev.example.com/tcp/30333/ws/p2p/<PEER_ID>
you'll see:
  Discovered new external address for our node:
    /dns/gdev.example.com/tcp/30333/ws/p2p/<PEER_ID>/p2p/<PEER_ID>
which is erroneous obviously.